backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
opensearch-py==2.4.2
cohere==4.47
sentence-transformers==2.5.1
pydantic==2.6.1
python-multipart==0.0.9
typing-extensions==4.9.0
requests==2.31.0
numpy==1.26.4
torch==2.2.0
transformers==4.37.2
pdfplumber==0.10.3
pandas==2.2.0
nltk==3.8.1
requests-aws4auth==1.2.3 